# Cache supplier for invocation rather than method resolving part 2