The second part with caching for methods, missed in the 1-st part

before
```
Benchmark                            Mode  Cnt     Score    Error   Units
JmhTest._bothifyExpression          thrpt    5  2730.742 ± 27.647  ops/ms
JmhTest._csv                        thrpt    5    29.307 ±  0.157  ops/ms
JmhTest._letterifyExpression        thrpt    5  2540.297 ± 31.841  ops/ms
JmhTest._optionsExpression          thrpt    5  4608.701 ± 81.775  ops/ms
JmhTest._regexifyExpression         thrpt    5  1645.380 ± 18.166  ops/ms
JmhTest.oneMethodCallExpression     thrpt    5  2212.006 ± 32.020  ops/ms
JmhTest.threeMethodsCallExpression  thrpt    5   523.795 ± 10.756  ops/ms
```

after
```
Benchmark                            Mode  Cnt     Score    Error   Units
JmhTest._bothifyExpression          thrpt    5  4277.270 ± 41.760  ops/ms
JmhTest._csv                        thrpt    5    35.568 ±  0.919  ops/ms
JmhTest._letterifyExpression        thrpt    5  4629.480 ± 62.134  ops/ms
JmhTest._optionsExpression          thrpt    5  4846.367 ± 34.680  ops/ms
JmhTest._regexifyExpression         thrpt    5  2070.759 ± 23.280  ops/ms
JmhTest.oneMethodCallExpression     thrpt    5  2309.779 ± 26.963  ops/ms
JmhTest.threeMethodsCallExpression  thrpt    5   565.426 ±  5.828  ops/ms
```